service add, remove & edit functionality

+++ CQUBS2/CQUBS2/src/java/ejb/ServiceAtLocationEJB.java
@@ -77,7 +77,21 @@ public class ServiceAtLocationEJB{
 
     //@Override
     public void deleteSAL(ServiceAtLocation sal) {
-        em.remove(em.merge(sal));
+        sal = em.find(ServiceAtLocation.class, sal.getSalId());
+        sal = em.merge(sal);
+        em.remove(sal);
+        /*Service service = em.find(Service.class, sal.getService().getServiceId());
+        
+        if(service.getSalList().size() == 1)
+        {
+            if(service.getSalList().get(0).getService().getServiceId() == sal.getSalId())
+            {
+                em.remove(em.merge(sal));
+                em.remove(em.merge(service));
+            }
+        }
+        else
+            em.remove(em.merge(sal));*/
     }
 
     //@Override
